+++ package.json
@@ -47,8 +47,10 @@
     "@testing-library/react": "^14.1.2",
     "@types/jest": "^29.5.11",
     "@types/lodash": "^4.17.7",
+    "@types/node": "^22.2.0",
     "@types/react": "^18.2.45",
     "@types/react-dom": "^18.2.18",
+    "@types/react-redux": "^7.1.33",
     "@typescript-eslint/eslint-plugin": "^6.12.0",
     "@typescript-eslint/parser": "^6.12.0",
     "commitizen": "^4.3.0",
@@ -62,18 +64,18 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "lint-staged": "^14.0.1",
-    "lodash": "^4.17.21",
     "mui-datatables": "^4.3.0",
     "notistack": "^3.0.1",
     "prettier": "^3.0.3",
     "prettier-plugin-organize-imports": "^3.2.3",
     "react-error-boundary": "^4.0.12",
     "react-markdown": "^8.0.7",
-    "react-redux": "^8.1.3",
+    "react-redux": "^8.1.1",
+    "redux": "^5.0.1",
     "rehype-raw": "^6.1.1",
     "remark-gfm": "^3.0.1",
     "ts-jest": "^29.1.1",
-    "tsup": "^8.0.1",
+    "tsup": "^8.2.4",
     "typescript": "^5.3.3"
   },
   "peerDependencies": {
@@ -82,12 +84,16 @@
     "@mui/material": "^5.15.11",
     "@types/mui-datatables": "*",
     "@xstate/react": "^4.1.1",
+    "lodash": "^4.17.21",
     "mui-datatables": "*",
     "react": ">=17",
     "react-dom": ">=17",
     "xstate": "^5.13.0"
   },
   "peerDependenciesMeta": {
+    "lodash": {
+      "optional": true
+    },
     "@emotion/react": {
       "optional": true
     },
