+++ src/redux-persist/index.ts
@@ -1,0 +1,2 @@
+export * from './PersistedStateProvider';
+export * from './initReduxPersist';
